refactor: remove next/image usage
