+++ Dockerfile
@@ -1,7 +1,5 @@
 FROM node:14-alpine
 
-# ENV RAZZLE_CUSTOM_VARIABLE x
-
 # install git, necessary for github dependencies
 RUN apk add --no-cache git
